@@ -3,7 +3,7 @@ description: Caror App
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -14,16 +14,19 @@ dependencies:
   http: ^0.13.4
   shared_preferences: ^2.0.13
   fluttertoast: 8.2.1
-  mobile_scanner: ^1.0.0
+  mobile_scanner: ^6.0.6
   flutter_rating_bar: ^4.0.0
   flutter_staggered_grid_view: ^0.4.1
   encrypt: ^5.0.1
   qr_flutter: ^4.0.0
   webview_flutter: ^3.0.2
-  audioplayers: ^0.20.1
+  audioplayers: ^6.2.0
   animations: ^2.0.2
   android_intent_plus: ^3.1.1
   flutter_native_splash: ^2.1.6
+  in_app_purchase: ^3.2.0
+  image_picker: ^1.1.2
+  share_plus: ^10.1.4
 
 dev_dependencies:
   flutter_test:
@@ -36,10 +39,13 @@ flutter:
   fonts:
     - family: Montserrat
       fonts:
-        - asset: fonts/Montserrat-Medium.ttf
-        - asset: fonts/Montserrat-SemiBold.ttf
+        - asset: assets/fonts/Montserrat-Medium.ttf
+        - asset: assets/fonts/Montserrat-SemiBold.ttf
           weight: 600
   assets:
     - assets/background_sound.mp3
+    - assets/app_icon.png
 flutter_intl:
   enabled: true
+  arb_dir: lib/resources/l10n
+  output_dir: lib/resources/generated
