@@ -1,4 +1,4 @@
-import 'package:caror/themes/number.dart';
+import 'package:caror/resources/number.dart';
 import 'package:caror/widget/widget.dart';
 import 'package:flutter/material.dart';
 import 'package:qr_flutter/qr_flutter.dart';
