@@ -1,8 +1,13 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
 class WebViewPage extends StatefulWidget {
-  const WebViewPage({Key? key}) : super(key: key);
+  const WebViewPage({
+    Key? key,
+    this.url = 'https://nguyenducthinh.com',
+  }) : super(key: key);
+
+  final String url;
 
   @override
   State<WebViewPage> createState() => _WebViewPageState();
@@ -23,8 +28,8 @@ class _WebViewPageState extends State<WebViewPage> {
           return true;
         },
         child: WebView(
-          backgroundColor: const Color(0xFF000000),
-          initialUrl: 'nguyenducthinh.info',
+          backgroundColor: Colors.white,
+          initialUrl: widget.url,
           javascriptMode: JavascriptMode.unrestricted,
           onWebViewCreated: (webViewController) {
             _controller = webViewController;
